requirements.txt
streamlit==1.32.0
pandas==2.1.1
numpy==1.26.0
plotly==5.18.0
yfinance==0.2.31
matplotlib==3.8.0
openpyxl==3.1.2
scipy==1.11.3
pillow==10.0.1
xlrd==2.0.1
requests==2.31.0
python-dotenv==1.0.0
streamlit-highcharts==0.2.0
git+https://github.com/Persevera-Asset-Management/PerseveraTools.git 